Parse contract
