+++ lib/src/contract/index.dart
@@ -1,2 +1,4 @@
+export 'model/index.dart';
 export 'contract.dart';
 export 'erc20.dart';
+export 'parse.dart';
