+++ lib/src/contract/model/index.dart
@@ -1,0 +1,3 @@
+export 'compiled_contract.dart';
+export 'entry_points_by_type.dart';
+export 'contract_abi.dart';
