+++ lib/src/declare.dart
@@ -1,0 +1,3 @@
+String compressProgram() {
+  return "";
+}
