+++ lib/starknet.dart
@@ -12,3 +12,4 @@ export 'src/types/index.dart';
 export 'src/convert.dart';
 export 'src/util.dart';
 export 'src/contract/index.dart';
+export 'src/contract/index.dart';
